# customer-service/src/main/java/com/nttdata/customerservice/api/dto/CustomerResponseDto.java
package com.nttdata.customerservice.api.dto;

import lombok.Builder;
import lombok.Data;

@Data
@Builder
public class CustomerResponseDto {

    private String customerId;
    private String firstName;
    private String lastName;
    private String documentType;
    private String documentNumber;
    private String email;
    private String phone;
    private String status;
    private String createdBy;
}