# customer-service/target/generated-sources/openapi/src/main/java/com/nttdata/customerservice/generated/api/CustomerApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.nttdata.customerservice.generated.api;

import com.nttdata.customerservice.generated.model.CreateCustomerRequest;
import com.nttdata.customerservice.generated.model.CustomerResponse;
import com.nttdata.customerservice.generated.model.UpdateCustomerRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import org.springframework.http.codec.multipart.Part;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-27T16:22:09.449758600-05:00[America/Lima]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Customer", description = "the Customer API")
public interface CustomerApi {

    default CustomerApiDelegate getDelegate() {
        return new CustomerApiDelegate() {};
    }

    /**
     * POST /api/customers : Create customer
     *
     * @param createCustomerRequest  (required)
     * @return Customer created (status code 201)
     *         or Invalid request (status code 400)
     *         or Customer already exists (status code 409)
     */
    @Operation(
        operationId = "createCustomer",
        summary = "Create customer",
        tags = { "Customer" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Customer created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CustomerResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request"),
            @ApiResponse(responseCode = "409", description = "Customer already exists")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/customers",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<CustomerResponse>> createCustomer(
        @Parameter(name = "CreateCustomerRequest", description = "", required = true) @Valid @RequestBody Mono<CreateCustomerRequest> createCustomerRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().createCustomer(createCustomerRequest, exchange);
    }


    /**
     * DELETE /api/customers/{customerId} : Delete customer
     *
     * @param customerId  (required)
     * @return Customer deleted (status code 204)
     *         or Customer not found (status code 404)
     */
    @Operation(
        operationId = "deleteCustomer",
        summary = "Delete customer",
        tags = { "Customer" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Customer deleted"),
            @ApiResponse(responseCode = "404", description = "Customer not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/customers/{customerId}"
    )
    
    default Mono<ResponseEntity<Void>> deleteCustomer(
        @Parameter(name = "customerId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("customerId") String customerId,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().deleteCustomer(customerId, exchange);
    }


    /**
     * GET /api/customers/{customerId} : Get customer by id
     *
     * @param customerId  (required)
     * @return Customer found (status code 200)
     *         or Customer not found (status code 404)
     */
    @Operation(
        operationId = "getCustomerById",
        summary = "Get customer by id",
        tags = { "Customer" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Customer found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CustomerResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Customer not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/customers/{customerId}",
        produces = { "application/json" }
    )
    
    default Mono<ResponseEntity<CustomerResponse>> getCustomerById(
        @Parameter(name = "customerId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("customerId") String customerId,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().getCustomerById(customerId, exchange);
    }


    /**
     * GET /api/customers : List customers
     *
     * @param documentNumber  (optional)
     * @return Customers retrieved (status code 200)
     */
    @Operation(
        operationId = "getCustomers",
        summary = "List customers",
        tags = { "Customer" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Customers retrieved", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = CustomerResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/customers",
        produces = { "application/json" }
    )
    
    default Mono<ResponseEntity<Flux<CustomerResponse>>> getCustomers(
        @Parameter(name = "documentNumber", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "documentNumber", required = false) String documentNumber,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().getCustomers(documentNumber, exchange);
    }


    /**
     * PUT /api/customers/{customerId} : Update customer
     *
     * @param customerId  (required)
     * @param updateCustomerRequest  (required)
     * @return Customer updated (status code 200)
     *         or Customer not found (status code 404)
     */
    @Operation(
        operationId = "updateCustomer",
        summary = "Update customer",
        tags = { "Customer" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Customer updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CustomerResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Customer not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/customers/{customerId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<CustomerResponse>> updateCustomer(
        @Parameter(name = "customerId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("customerId") String customerId,
        @Parameter(name = "UpdateCustomerRequest", description = "", required = true) @Valid @RequestBody Mono<UpdateCustomerRequest> updateCustomerRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().updateCustomer(customerId, updateCustomerRequest, exchange);
    }

}